Move Dashboard to pages
Move Dashboard page to pages folder as Dashboard is rightfully a page.
MatchCard however is a UI component. Dashboard sections such as
DashboardCardsSection, DashboardSidebar, MatchesSection and Profile are layout
specific components that are not reusable outside of the Dashboard page.
