+++ imports/ui/containers/dashboard.jsx
@@ -1,7 +1,6 @@
-import { Meteor } from 'meteor/meteor';
 import { withTracker } from 'meteor/react-meteor-data';
 import Entries from '/imports/api/entries/entries.js';
-import Dashboard from '/imports/ui/components/dashboard/dashboard.jsx';
+import Dashboard from '/imports/ui/pages/dashboard/dashboard';
 
 // eslint-disable-next-line no-undef
 const categoryId = new ReactiveVar(3);
