+++ imports/ui/pages/dashboard/dashboard.jsx
@@ -1,9 +1,11 @@
 import React, { Component } from 'react';
+import { PropTypes } from 'prop-types';
+
+import { categories } from '/lib/data/categories.js';
+
 import DashboardSidebar from './dashboard_sidebar';
 import MatchesSection from './matches_section';
 import DashboardCardsSection from './dashboard_cards_section';
-import { categories } from '/lib/data/categories.js';
-import { PropTypes } from 'prop-types';
 
 const sectionTargets = {
   //see: categories.js for numbers
@@ -15,6 +17,7 @@ const sectionTargets = {
   'accountability-partner': 1,
   other: 2,
 };
+
 class Dashboard extends Component {
   constructor(props) {
     super(props);
